requirements.txt
mss>=9.0.0
PyQt5>=5.15.0
pystray>=0.19.0
pynput>=1.7.0
Pillow>=10.0.0
pywin32>=306
